Clean compile on Linux
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1950 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/wp64/kernel/parser/SourceFile.cpp
@@ -4774,7 +4774,7 @@ void RexxSource::parseTraceSetting(
 
         case '?':                      /* debug toggle character            */
                                        /* already toggling?                 */
-          if (*debug == DEBUG_TOGGLE)
+          if (*debug == (size_t)DEBUG_TOGGLE)
             *debug = DEBUG_IGNORE;     /* this is back to no change at all  */
           else
             *debug = DEBUG_TOGGLE;     /* need to toggle the debug mode     */
